requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mangum==0.18.0
python-telegram-bot==20.5
asyncpg==0.29.0
python-dotenv==1.0.1
